New Additions: Added code for automatically matching compatible listings <- didn't write this but this is pretty much it I think.
Also, Josh L fixed the bug with caller because I can't do the polymorphing.

Problems:
It seems indeterminate whether the search is querying/finding the data to match the query.  Even if I enter something like "Crabs Snakes Beer" it will say it's in the database, when it is in fact not.  So we'll have to look into that.